Rework query tags display
